backend/requirements.txt
fastapi
uvicorn
python-multipart
# AI Pipeline
langchain
langchain-core
langchain-openai
langchain-community
# Vector Database
chromadb
# Document Loaders & extraction
pypdf
python-docx
python-dotenv
